+++ Changelog.md
@@ -1,5 +1,7 @@
 ﻿# v$vNext$ - $buildDate$
 
+* Feature: (#11) Upadte binding registry after define step
+
 # v2021.1.4 - 2021-11-25
 
 * Fix: (#29) Disable postbuild events from connectors